pyproject.toml
# https://rudeigerc.dev/posts/structuring-a-python-project/
[project]
name = "python-template"
dynamic = ["version"]
description = "Default template for PDM package"
authors = [
    {name = "Marco Berghoff", email = "marco.berghoff@kit.edu"},
]
dependencies = ["click"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]

[project.scripts]
greet = "python_template.__main__:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[dependency-groups]
lint = [
    "ruff>=0.8.1",
    "pylint>=3.3.1",
    "mypy>=1.13.0",
]
pytest = [
    "pytest-cov>=6.0.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.47",
]

[tool.ruff]
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]
line-length = 120

[tool.ruff.lint]
select = [
    "ARG",
    "ANN",  # Warnungen für fehlende Typannotationen
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "D",   # Missing docstrings (autodoc)
    "DOC", # pydoclint
    "E", # pycodestyle - Error
    "F", # Pyflakes
    "I", # isort
    "W", # pycodestyle - Warning
    "UP", # pyupgrade
    #"PL", # Pylint
]
extend-select = ["DOC"]
preview = true
ignore = []

[tool.ruff.lint.per-file-ignores] 
"tests/**" = ["PLC2701"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pylint]
load-plugins=["pylint.extensions.docparams"]
[tool.pylint.main]
ignore-paths=[
    ".venv",
    ".git",
    "__pypackages__",
    "tests/_temp",
    "tests"
]
accept-no-param-doc="no"
accept-no-raise-doc="no"
accept-no-yields-doc="no"
default-docstring-type="google"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
lint = "ruff check ."
fmt = "ruff format ."
docs = "mkdocs serve"

[tool.pdm.version]
source = "scm"

[tool.mypy]
strict = true

